Work in progress. Not all tests are passing yet.
Add a DestContext parameter to the clone methods of the debug info objects derived from MDNode. Clone metadata across contexts.

Add RF_LinkAcrossContexts to ValueMapper's RemapFlags so it knows whether to remap across contexts. Add a SrcContext and DestContext to it too.

Add ValueMapper::mapAttributeList. Technically attribute lists aren't values but we need to do this during scheduled remappings so it has to live on the ValueMapper. Moved from IRLinker. Add "cloneIntoContext" that clones an Attribute into the new context.

Improve errors from the verifier when attributes are in the wrong context.

Add "remap global object metadata" as a new type of action that can be scheduled on the value mapper. Only when relinking across contexts do we need to change the metadata attachment kind itself, not just the MDNode it's attaching to. Decreases #bits available for MCID from 29 to 28, hope that's ok?

Load the syncscope names and metadata kind names from the source context when the Mapper is constructed instead of lazily.